+++ src/main/kotlin/dev/codebandits/container/gradle/tasks/ExecutionStep.kt
@@ -1,7 +1,5 @@
 package dev.codebandits.container.gradle.tasks
 
-import org.gradle.api.Action
-import org.gradle.api.Task
 import org.gradle.process.ExecResult
 
 public class ExecutionStep(