+++ src/testFeatures/kotlin/dev/codebandits/container/gradle/GroovySyntaxTest.kt
@@ -26,7 +26,7 @@ class GroovySyntaxTest : GradleProjectTest() {
     buildGradleFile.appendLine(
       """
       import dev.codebandits.container.gradle.tasks.ContainerTask
-      import dev.codebandits.container.gradle.tasks.outputLocalImage
+      import dev.codebandits.container.gradle.image.outputLocalImage
       
       plugins {
         id("dev.codebandits.container")