+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Chore
 
-- Simplify image input/output API.
+- Improve image input/output API.
 
 ## [0.2.1] - 2025-04-23